fixed hw 2 using floats instead of doubles

+++ MyHW/Hw problem 2/main.cpp
@@ -1,6 +1,6 @@
 /* 
  * File:   main.cpp
- * Author: ricky
+ * Author: Ricardo Luna
  *
  * Created on February 26, 2022, 8:31 PM
  */
@@ -11,10 +11,10 @@ using namespace std;
 
 int main() { 
     // storing integers total sales and revenue 
-    double tsales = .58;
-    double revenue = 8600000;
+    float tsales = .58;
+    float revenue = 8600000;
     
-    int total = revenue / tsales;
+    float total = revenue / tsales;
     
     cout << total;
